// libs/supabase/schema.ts
export interface Database {
// ...
  public: {
    Tables: {
      gists: {
        Row: {
          content: string
          created_at: string
          description: string
          id: string
          is_paid: boolean
          lang: string
          price: number
          profile_uid: string
          title: string
        }
        Insert: {
          content: string
          created_at?: string
          description: string
          id: string
          is_paid: boolean
          lang: string
          price: number
          profile_uid: string
          title: string
        }
        Update: {
          content?: string
          created_at?: string
          description?: string
          id?: string
          is_paid?: boolean
          lang?: string
          price?: number
          profile_uid?: string
          title?: string
        }
        Relationships: [
          {
            foreignKeyName: 'gists_profile_uid_fkey'
            columns: ['profile_uid']
            isOneToOne: false
            referencedRelation: 'profiles'
            referencedColumns: ['id']
          },
        ]
      }
// ...
    }
    Views: {
      [_ in never]: never;
    }
    Functions: {
      [_ in never]: never;
    }
    Enums: {
      [_ in never]: never;
    }
    CompositeTypes: {
      [_ in never]: never;
    }
  }
// ...
}